pyproject.toml
[tool.poetry]
name = "cbs-lessons"
version = "0.1.0"
description = ""
authors = ["umterrick <usichenko.vlad67@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pytest = "^8.2.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


